refactor: centralize badge sorting and claim status logic in BadgeProvider
